convert px to sp and fixed bug

+++ app/src/main/java/com/blueapps/egyptianwriter/editor/document/properties/PropertiesFragment.java
@@ -53,7 +53,7 @@ public class PropertiesFragment extends Fragment implements TextWatcher {
         textSizeDecrease = binding.textSizeDecrease;
 
         // Set initial values
-        editTextSize.setText((String.valueOf(TypedValueCompat.pxToSp(propertiesManager.getTextSize().getValue(), getResources().getDisplayMetrics()))));
+        editTextSize.setText((String.valueOf((int) TypedValueCompat.pxToSp(propertiesManager.getTextSize().getValue(), getResources().getDisplayMetrics()))));
 
         // textSize
         Handler handler = new Handler(Looper.getMainLooper());
